+++ visualization/app/codeCharta/state/effects/autoFitCodeMapChange/autoFitCodeMap.effect.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core"
 import { Store } from "@ngrx/store"
-import { createEffect } from "@ngrx/effects"
-import { switchMap, filter, skip, take, tap, combineLatest, withLatestFrom } from "rxjs"
+import { Actions, createEffect, ofType } from "@ngrx/effects"
+import { switchMap, filter, skip, take, tap, combineLatest, withLatestFrom, first } from "rxjs"
 import { CcState } from "../../../codeCharta.model"
 import { ThreeMapControlsService } from "../../../ui/codeMap/threeViewer/threeMapControls.service"
 import { visibleFileStatesSelector } from "../../selectors/visibleFileStates.selector"
@@ -15,7 +15,8 @@ export class AutoFitCodeMapEffect {
     constructor(
         private store: Store<CcState>,
         private renderCodeMapEffect: RenderCodeMapEffect,
-        private threeMapControlsService: ThreeMapControlsService
+        private threeMapControlsService: ThreeMapControlsService,
+        private actions$: Actions
     ) {}
 
     autoFitTo$ = createEffect(
@@ -28,6 +29,19 @@ export class AutoFitCodeMapEffect {
                 skip(1), // initial map load is already fitted
                 withLatestFrom(this.store.select(resetCameraIfNewFileIsLoadedSelector)),
                 filter(([, resetCameraIfNewFileIsLoaded]) => resetCameraIfNewFileIsLoaded),
+                switchMap(() => this.renderCodeMapEffect.renderCodeMap$.pipe(take(1))),
+                tap(() => {
+                    this.threeMapControlsService.autoFitTo()
+                })
+            ),
+        { dispatch: false }
+    )
+
+    autoFitToWhenResetCameraIfNewFileIsLoadedSetToFalse$ = createEffect(
+        () =>
+            this.actions$.pipe(
+                ofType("StartWithGlobalOption:resetCameraIfNewFileIsLoadedSetToFalse"),
+                first(),
                 switchMap(() => this.renderCodeMapEffect.renderCodeMap$.pipe(take(1))),
                 tap(() => {
                     this.threeMapControlsService.autoFitTo()